Split API test helpers
Move exec-specific API tests into their own submodule and extract the
experimental log-capture fixture from `src/api/tests.rs`. Keep the main
API test module under the repository file-size cap without changing the
covered behaviours.
