+++ src/api/tests.rs
@@ -1,41 +1,25 @@
 //! Unit tests for the orchestration API module.
 
-use bollard::container::LogOutput;
-use futures_util::stream;
-use mockall::mock;
 use proptest::prelude::*;
 use rstest::rstest;
 #[cfg(feature = "experimental")]
-use std::sync::{
-    Arc, Mutex,
-    atomic::{AtomicUsize, Ordering},
-};
+use std::sync::atomic::{AtomicUsize, Ordering};
 
-use super::exec::exec_with_client;
-use super::{CommandOutcome, ExecMode, ExecRequest, RunRequest};
+use super::{CommandOutcome, RunRequest};
 #[cfg(feature = "experimental")]
 use super::{list_containers, run_agent, run_token_daemon, stop_container};
 #[cfg(feature = "experimental")]
 use crate::config::{AppConfig, GitHubConfig};
-use crate::engine::{
-    ContainerExecClient, CreateExecFuture, ExecMode as EngineExecMode, InspectExecFuture,
-    ResizeExecFuture, StartExecFuture,
-};
+#[cfg(feature = "experimental")]
 use crate::error::{ConfigError, PodbotError};
 #[cfg(feature = "experimental")]
 use camino::Utf8PathBuf;
 
-mock! {
-    #[derive(Debug)]
-    ApiExecClient {}
-
-    impl ContainerExecClient for ApiExecClient {
-        fn create_exec(&self, container_id: &str, options: bollard::exec::CreateExecOptions<String>) -> CreateExecFuture<'_>;
-        fn start_exec(&self, exec_id: &str, options: Option<bollard::exec::StartExecOptions>) -> StartExecFuture<'_>;
-        fn inspect_exec(&self, exec_id: &str) -> InspectExecFuture<'_>;
-        fn resize_exec(&self, exec_id: &str, options: bollard::exec::ResizeExecOptions) -> ResizeExecFuture<'_>;
-    }
-}
+mod exec;
+#[cfg(feature = "experimental")]
+mod log_capture;
+#[cfg(feature = "experimental")]
+use log_capture::{capture_logs, capture_warning_logs, require_log_contains, require_outcome};
 
 #[rstest]
 fn command_outcome_success_equals_itself() {
@@ -108,121 +92,6 @@ proptest! {
             prop_assert_eq!(request.branch(), branch);
         }
     }
-}
-
-#[rstest]
-fn exec_request_defaults_to_attached_without_tty() {
-    let request =
-        ExecRequest::new("sandbox", vec![String::from("echo")]).expect("request should be valid");
-
-    assert_eq!(request.mode(), ExecMode::Attached);
-    assert!(!request.tty());
-}
-
-#[rstest]
-#[case(ExecMode::Attached, EngineExecMode::Attached)]
-#[case(ExecMode::Detached, EngineExecMode::Detached)]
-#[case(ExecMode::Protocol, EngineExecMode::Protocol)]
-fn exec_mode_maps_to_engine_mode(
-    #[case] api_mode: ExecMode,
-    #[case] expected_engine_mode: EngineExecMode,
-) {
-    let engine_mode: EngineExecMode = api_mode.into();
-    assert_eq!(engine_mode, expected_engine_mode);
-}
-
-#[rstest]
-fn exec_request_builder_methods_preserve_other_fields() {
-    let base = ExecRequest::new("sandbox", vec![String::from("echo"), String::from("hello")])
-        .expect("request should be valid");
-
-    let updated = base.clone().with_mode(ExecMode::Protocol).with_tty(true);
-
-    assert_eq!(updated.container(), "sandbox");
-    assert_eq!(
-        updated.command(),
-        &[String::from("echo"), String::from("hello")]
-    );
-    assert_eq!(updated.mode(), ExecMode::Protocol);
-    assert!(!updated.tty());
-
-    assert_eq!(base.container(), "sandbox");
-    assert_eq!(
-        base.command(),
-        &[String::from("echo"), String::from("hello")]
-    );
-    assert_eq!(base.mode(), ExecMode::Attached);
-    assert!(!base.tty());
-}
-
-#[rstest]
-#[case(ExecMode::Detached)]
-#[case(ExecMode::Protocol)]
-fn exec_request_normalizes_tty_for_non_attached_modes(#[case] mode: ExecMode) {
-    let request = ExecRequest::new("sandbox", vec![String::from("echo")])
-        .expect("request should be valid")
-        .with_tty(true)
-        .with_mode(mode)
-        .with_tty(true);
-
-    assert_eq!(request.mode(), mode);
-    assert!(
-        !request.tty(),
-        "tty should be disabled for non-attached modes"
-    );
-}
-
-#[rstest]
-fn exec_request_rejects_blank_container() {
-    let error = ExecRequest::new("   ", vec![String::from("echo")])
-        .expect_err("blank container should be rejected");
-
-    assert!(matches!(
-        error,
-        PodbotError::Config(ConfigError::MissingRequired { field }) if field == "container"
-    ));
-}
-
-#[rstest]
-fn exec_request_rejects_empty_command() {
-    let error =
-        ExecRequest::new("sandbox", Vec::new()).expect_err("empty command should be rejected");
-
-    assert!(matches!(
-        error,
-        PodbotError::Config(ConfigError::MissingRequired { field }) if field == "command"
-    ));
-}
-
-#[rstest]
-fn exec_request_rejects_blank_executable() {
-    let error = ExecRequest::new("sandbox", vec![String::from("  ")])
-        .expect_err("blank executable should be rejected");
-
-    assert!(matches!(
-        error,
-        PodbotError::Config(ConfigError::MissingRequired { field }) if field == "command[0]"
-    ));
-}
-
-#[rstest]
-#[case::zero_exit_code(0, CommandOutcome::Success)]
-#[case::non_zero_exit_code(42, CommandOutcome::CommandExit { code: 42 })]
-fn exec_with_client_maps_exit_code_to_outcome(
-    #[case] exit_code: i64,
-    #[case] expected: CommandOutcome,
-) {
-    let base_request = ExecRequest::new("sandbox", vec![String::from("echo"), String::from("ok")])
-        .expect("request should be valid");
-    let request = base_request.with_mode(ExecMode::Detached);
-    let runtime = tokio::runtime::Runtime::new().expect("runtime should be created");
-    let mut client = MockApiExecClient::new();
-    configure_exec_client(&mut client, request.mode(), exit_code);
-
-    let outcome = exec_with_client(&client, runtime.handle(), &request)
-        .expect("exit code should map to a command outcome");
-
-    assert_eq!(outcome, expected);
 }
 
 #[rstest]
@@ -348,110 +217,6 @@ fn credential_validation_thread_panic_maps_to_github_error() {
     ));
 }
 
-#[cfg(feature = "experimental")]
-#[derive(Clone)]
-struct SharedLogWriter {
-    output: Arc<Mutex<Vec<u8>>>,
-}
-
-#[cfg(feature = "experimental")]
-struct SharedLogBuffer {
-    output: Arc<Mutex<Vec<u8>>>,
-}
-
-#[cfg(feature = "experimental")]
-impl<'writer> tracing_subscriber::fmt::MakeWriter<'writer> for SharedLogWriter {
-    type Writer = SharedLogBuffer;
-
-    fn make_writer(&'writer self) -> Self::Writer {
-        SharedLogBuffer {
-            output: Arc::clone(&self.output),
-        }
-    }
-}
-
-#[cfg(feature = "experimental")]
-impl std::io::Write for SharedLogBuffer {
-    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-        let mut output = self
-            .output
-            .lock()
-            .map_err(|_| std::io::Error::other("log buffer mutex poisoned"))?;
-        output.extend_from_slice(buf);
-        Ok(buf.len())
-    }
-
-    fn flush(&mut self) -> std::io::Result<()> {
-        Ok(())
-    }
-}
-
-#[cfg(feature = "experimental")]
-fn capture_warning_logs(
-    operation: impl FnOnce(),
-) -> Result<String, Box<dyn std::error::Error + Send + Sync>> {
-    capture_logs(operation, tracing::Level::WARN)
-}
-
-#[cfg(feature = "experimental")]
-fn capture_logs(
-    operation: impl FnOnce(),
-    max_level: tracing::Level,
-) -> Result<String, Box<dyn std::error::Error + Send + Sync>> {
-    let output = Arc::new(Mutex::new(Vec::new()));
-    let writer = SharedLogWriter {
-        output: Arc::clone(&output),
-    };
-    let subscriber = tracing_subscriber::fmt()
-        .with_writer(writer)
-        .with_ansi(false)
-        .without_time()
-        .with_max_level(max_level)
-        .finish();
-
-    tracing::subscriber::with_default(subscriber, operation);
-
-    let bytes = output
-        .lock()
-        .map_err(|error| {
-            Box::new(std::io::Error::other(format!(
-                "log buffer mutex poisoned: {error}"
-            ))) as Box<dyn std::error::Error + Send + Sync>
-        })?
-        .clone();
-    String::from_utf8(bytes)
-        .map_err(|error| Box::new(error) as Box<dyn std::error::Error + Send + Sync>)
-}
-
-#[cfg(feature = "experimental")]
-fn require_outcome(
-    actual: CommandOutcome,
-    expected: CommandOutcome,
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    if actual == expected {
-        Ok(())
-    } else {
-        Err(Box::new(std::io::Error::other(format!(
-            "expected outcome {expected:?}, got {actual:?}"
-        ))))
-    }
-}
-
-#[cfg(feature = "experimental")]
-fn require_log_contains(
-    logs: &str,
-    expected: &str,
-    description: &str,
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    if logs.contains(expected) {
-        Ok(())
-    } else {
-        Err(Box::new(std::io::Error::other(format!(
-            "warning should include {description}: {logs}"
-        ))))
-    }
-}
-
 #[rstest]
 #[cfg(feature = "experimental")]
 fn credential_validation_uses_local_runtime_without_current_handle() {
@@ -508,45 +273,6 @@ fn credential_validation_scoped_thread_panic_maps_to_github_error() {
 }
 
 #[rstest]
-#[case(r#"{"container":"   ","command":["echo"]}"#, "container")]
-#[case(r#"{"container":"sandbox","command":[]}"#, "command")]
-#[case(r#"{"container":"sandbox","command":["   "]}"#, "command[0]")]
-fn exec_request_deserialization_reuses_validation(
-    #[case] payload: &str,
-    #[case] expected_field: &str,
-) {
-    let error = serde_json::from_str::<ExecRequest>(payload)
-        .expect_err("invalid payload should fail validation");
-
-    assert!(
-        error.to_string().contains(expected_field),
-        "expected error to mention {expected_field}, got: {error}"
-    );
-}
-
-#[rstest]
-#[case(
-    r#"{"container":"sandbox","command":["echo"],"mode":"Detached","tty":true}"#,
-    ExecMode::Detached
-)]
-#[case(
-    r#"{"container":"sandbox","command":["echo"],"mode":"Protocol","tty":true}"#,
-    ExecMode::Protocol
-)]
-fn exec_request_deserialization_normalizes_tty_for_non_attached_modes(
-    #[case] payload: &str,
-    #[case] expected_mode: ExecMode,
-) {
-    let request = serde_json::from_str::<ExecRequest>(payload).expect("payload should deserialize");
-
-    assert_eq!(request.mode(), expected_mode);
-    assert!(
-        !request.tty(),
-        "tty should be disabled for non-attached modes"
-    );
-}
-
-#[rstest]
 #[case::run_agent("run_agent")]
 #[case::list_containers("list_containers")]
 #[case::stop_container("stop_container")]
@@ -564,56 +290,4 @@ fn stub_returns_success(#[case] stub: &str) {
     }
     .expect("stub should return Ok");
     assert_eq!(outcome, CommandOutcome::Success);
-}
-
-fn configure_exec_client(client: &mut MockApiExecClient, mode: ExecMode, exit_code: i64) {
-    client.expect_create_exec().times(1).returning(|_, _| {
-        Box::pin(async {
-            Ok(bollard::exec::CreateExecResults {
-                id: String::from("api-exec-id"),
-            })
-        })
-    });
-
-    match mode {
-        ExecMode::Attached | ExecMode::Protocol => {
-            client.expect_start_exec().times(1).returning(move |_, _| {
-                let output_stream = stream::iter(vec![Ok(LogOutput::StdOut {
-                    message: Vec::from(&b"api output"[..]).into(),
-                })]);
-                Box::pin(async move {
-                    Ok(bollard::exec::StartExecResults::Attached {
-                        output: Box::pin(output_stream),
-                        input: Box::pin(tokio::io::sink()),
-                    })
-                })
-            });
-        }
-        ExecMode::Detached => {
-            client.expect_start_exec().times(1).returning(|_, _| {
-                Box::pin(async { Ok(bollard::exec::StartExecResults::Detached) })
-            });
-        }
-    }
-
-    match mode {
-        ExecMode::Attached => {
-            client
-                .expect_resize_exec()
-                .times(0..)
-                .returning(|_, _| Box::pin(async { Ok(()) }));
-        }
-        ExecMode::Detached | ExecMode::Protocol => {
-            client.expect_resize_exec().never();
-        }
-    }
-
-    client.expect_inspect_exec().times(1).returning(move |_| {
-        let inspect = bollard::models::ExecInspectResponse {
-            running: Some(false),
-            exit_code: Some(exit_code),
-            ..bollard::models::ExecInspectResponse::default()
-        };
-        Box::pin(async move { Ok(inspect) })
-    });
 }
